refactor: LOCAL_STORAGE_KEY -> STORAGE_KEY 로 변경
