+++ hooks/useSearchForm.ts
@@ -1,6 +1,6 @@
 import { ChangeEvent, useState } from 'react';
 import { getLocalStorageValue, setLocalStorageValue } from '../shared/utils/localStorage';
-import { LOCAL_STORAGE_KEY } from '../shared/constants/localStorageKey';
+import { LOCAL_STORAGE_KEY } from '../shared/constants/storageKey';
 import { useRouter } from 'next/router';
 import useToast from './useToast';
 import { ToastType } from '../shared/type/common';
