+++ shared/type/localStorage.ts
@@ -1,4 +1,3 @@
-import { LOCAL_STORAGE_KEY } from '../constants/localStorageKey';
+import { LOCAL_STORAGE_KEY } from '../constants/storageKey';
 
-export type LocalStorageKey =
-  typeof LOCAL_STORAGE_KEY[keyof typeof LOCAL_STORAGE_KEY];
+export type LocalStorageKey = typeof LOCAL_STORAGE_KEY[keyof typeof LOCAL_STORAGE_KEY];
